+++ src/ext/const.py
@@ -9,17 +9,22 @@ from datetime import datetime
 import discord
 from discord.ext import commands
 from ext import excepts
-from ext.logcfg import get_logger, logging
+from ext.logcfg import gLogr
 from modules.consolemod import style
+__all__ = [
+    "fix_settings", "DEFAULT_SETTINGS", "chk_sudo", "log", "bot", "get_prefix",
+    "STATUSES", "BOTSETFILE", "LASTWRDFILE", "SETFILE", "WARNFILE", "STRFILE", "TAGFILE"
+]
 
 
 STATUSES = [
-    '2020 Best discord bot: Merlin', 'PyPI', 'Github', 'Repl.it', 'Minecraft', 'Windows Whistler OOBE', 'GitLab', 'readthedocs.io', 'NoCopyrightSounds', 'Discord',
+    'Best discord bot: Merlin', 'PyPI', 'Github', 'Repl.it', 'Minecraft', 'Windows Whistler OOBE', 'GitLab', 'readthedocs.io', 'NoCopyrightSounds', 'Discord',
     'Recursion', 'F0rk B0mbs', 'Different 𝗞𝗶𝗻𝗱𝘀 𝘖𝘧 𝙲𝚑𝚊𝚛𝚊𝚌𝚝𝚎𝚛𝚜', 'sudo rm -rf / --no-preserve-root', 'rd/s/q %windir%', 'typing "exit" in linux init=/bin/bash',
     'Hello, world!', 'Oracle Virtualbox VMs', 'VMware', 'Quick EMUlator (QEMU)', 'Global Information Tracker', 'Goddamn Idiotic Truckload of sh*t',
     'Arch Linux', 'Manjaro Linux', 'Microsoft Windows 10', 'Canonical Ubuntu', 'Kubuntu and Xubuntu', 'Linux Mint', 'Pop!_OS', 'OpenSUSE', 'Elementry OS', 'MX Linux', 'Debian', 'BSD',
     'Nothing', 'Status', 'what Merlin is playing', 'Twitter', 'StackOverflow', 'Mozilla Firefox', 'Visual Studio Code', 'zsh', 'fish', 'dash', 'mc (Midnight Commander)',
-    'Ruby On Rails', 'Python', 'JavaScript', 'Node.js', 'Angular', 'Assembly', 'C++ (see ga ga)', 'C', 'Docker', 'Java', 'ps1', 'Nim', 'Markdown', 'HTML', 'CSS', 'Perl', 'C#', 'R', 'Pascal']
+    'Ruby On Rails', 'Python', 'JavaScript', 'Node.js', 'Angular', 'Assembly', 'C++ (see ga ga)', 'C', 'Docker', 'Java', 'ps1', 'Nim', 'Markdown', 'HTML', 'CSS', 'Perl', 'C#', 'R', 'Pascal'
+]
 
 
 # path for file storing data
@@ -31,47 +36,10 @@ STRFILE = "ext/wrds.json"
 BOTSETFILE = "ext/bot_settings.json"
 SETFILE = "data/settings.json"
 TAGFILE = "data/tags.db"
+RANKFILE = "data/rank.db"
 
 
-logger, eventLogger, cmdHdlLogger = get_logger('Merlin'), get_logger('EVENT'), get_logger('CMDHDL')
-logging.basicConfig(filename='discordbot.log', level=15, format='[%(asctime)s]%(levelname)s - %(name)s: %(message)s')
-HINT_LEVEL_NUM = 17
-logging.addLevelName(HINT_LEVEL_NUM, "HINT")
-
-
-def hint(self, message, *args, **kws):
-    """hint logging level"""
-    if self.isEnabledFor(HINT_LEVEL_NUM):
-        # Yes, logger takes its '*args' as 'args'.
-        self._log(HINT_LEVEL_NUM, message, args, **kws)
-
-
-def slog(message: str):
-    """sub log"""
-    print(' >> ' + message)
-    logger.hint(message)
-
-
-def nlog(message: str):
-    """new line long"""
-    print('\n==> ' + message)
-    logger.info(message)
-
-
-def cmd_handle_log(message: str):
-    """logging function for command handling"""
-    print('[CMDHDL]\t' + message)
-    cmdHdlLogger.info(message)
-
-
-def event_log(message: str):
-    print('[EVENT]\t' + message)
-    eventLogger.info(message)
-
-
-def cmd_handle_warn(message: str):
-    print(style.orange + message + style.reset)
-    cmdHdlLogger.warning(message)
+logger, eventLogger, cmdHdlLogger = gLogr('Merlin.root', 'Merlin.event', 'Merlin.cmdHdl')
 
 
 def get_prefix(bot: commands.Bot, message: discord.Message):
@@ -93,14 +61,9 @@ def get_prefix(bot: commands.Bot, message: discord.Message):
         return tuple(prefixes)
 
 
-bot = discord.ext.commands.Bot(
-    command_prefix=get_prefix,
-    description="an awesome discord bot coded in discord.py",
-    owner_id=653086042752286730,
-    case_insensitive=True)
-
-
 class Log:
+    def __init__(self, bot):
+        self.bot = bot
     @staticmethod
     async def worker_log(name, queue):
         slept = 0
@@ -116,8 +79,7 @@ class Log:
             queue.task_done()
             slept = 0  # reset timeout
 
-    @classmethod
-    async def __call__(cls, message: str, *, guild: discord.Guild = None):
+    async def __call__(self, message: str, guild: discord.Guild = None):
         if guild:
             for channel in guild.channels:
                 if channel.name == 'merlin-py':
@@ -126,15 +88,12 @@ class Log:
         queue = asyncio.Queue()
         tasks = []
         for i in range(5):
-            tasks.append(asyncio.create_task(cls.worker_log(f'worker-log-{i}', queue)))
-        for guild in bot.guilds:
+            tasks.append(asyncio.create_task(self.worker_log(f'worker-log-{i}', queue)))
+        for guild in self.bot.guilds:
             for channel in guild.channels:
                 if channel.name == 'merlin-py':
                     queue.put_nowait((channel, message))
                     break
-
-
-log = Log()
 
 
 def is_sudoers(member: discord.Member):
@@ -178,29 +137,8 @@ DEFAULT_SETTINGS = {
     "cmdHdl": {
         "cmdNotFound": 0,
         "delIssue": 0,
-        "improveExp": 0}}
+        "improveExp": 0
+    }
+}
 
 
-def fix_settings(guild: discord.Guild):
-    settings = None
-    try:
-        settings = json.load(open(SETFILE, 'r'))
-    except Exception:
-        settings = {}
-        open(SETFILE, 'w').write("{}")
-    try:
-        settings[f"g{guild.id}"]['cmdHdl']
-    except KeyError:
-        settings[f"g{guild.id}"] = DEFAULT_SETTINGS.copy()
-        with open(SETFILE, 'w') as outfile:
-            json.dump(settings, outfile)
-        return
-    # fix cmdHdl
-    cmdHdl = DEFAULT_SETTINGS['cmdHdl'].copy()          # the following code will leave entrys already
-    cmdHdl.update(settings[f'g{guild.id}']['cmdHdl'])   # exists and add the missing entrys so that
-    settings[f'g{guild.id}']['cmdHdl'] = cmdHdl         # overwriting can be prevented
-    default = DEFAULT_SETTINGS.copy()
-    default.update(settings[f'g{guild.id}'])            # we can also do the same thing for the whole settings
-    settings[f'g{guild.id}'].update(default)
-    with open(SETFILE, 'w') as outfile:
-        json.dump(settings, outfile)